feat: currently reading functionalities added

+++ src/components/books/Book.tsx
@@ -1,12 +1,21 @@
 /* eslint-disable @typescript-eslint/no-unsafe-argument */
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import { Badge, Button, Card, Flex, Grid, Group, Text } from "@mantine/core"
-import { IconBookmarkMinus, IconBookmarkPlus } from "@tabler/icons-react"
+import {
+  IconBook,
+  IconBook2,
+  IconBookmarkMinus,
+  IconBookmarkPlus,
+} from "@tabler/icons-react"
 import { toast } from "react-hot-toast"
 import { useSelector } from "react-redux"
 import { Link } from "react-router-dom"
 import { selectUser } from "../../features/auth/authSlice"
 
+import {
+  useAddNewReadingMutation,
+  useDeleteReadingMutation,
+} from "../../features/reading/wishlistApi"
 import {
   useAddNewWishlistMutation,
   useDeleteWishlistMutation,
@@ -17,30 +26,63 @@ import { ErrorResponse, TBook } from "../../types"
 type Param = {
   book: TBook
   bookmarkId: { _id: string }
+  readingId: { _id: string }
 }
 
 export default function Book(param: Param) {
-  const { book, bookmarkId } = param
-  const [addWishlist] = useAddNewWishlistMutation()
-  const [deleteWishlist] = useDeleteWishlistMutation()
+  const { book, bookmarkId, readingId } = param
+  const [addWishlist, { isLoading: isaddWishlistLoading }] =
+    useAddNewWishlistMutation()
+  const [deleteWishlist, { isLoading: isdeleteWishlistLoading }] =
+    useDeleteWishlistMutation()
+  const [addReading, { isLoading: isaddReadingLoading }] =
+    useAddNewReadingMutation()
+  const [deleteReading, { isLoading: isdeleteReadingLoading }] =
+    useDeleteReadingMutation()
   const auth = useAuth()
   const user = useSelector(selectUser) as unknown as { email: string }
 
   const handleAddBookmark = async () => {
-    try {
-      await addWishlist({ bookId: book._id, email: user.email }).unwrap()
-      toast.success("Bookmark added successfully")
-    } catch (error) {
-      toast.error((error as ErrorResponse).data.message)
+    if (!isaddWishlistLoading) {
+      try {
+        await addWishlist({ bookId: book._id, email: user.email }).unwrap()
+        toast.success("Bookmark added successfully")
+      } catch (error) {
+        toast.error((error as ErrorResponse).data.message)
+      }
     }
   }
 
   const handleDeleteBookmark = async () => {
-    try {
-      await deleteWishlist(bookmarkId._id).unwrap()
-      toast.success("Bookmark deleted successfully")
-    } catch (error) {
-      toast.error((error as ErrorResponse).data.message)
+    if (!isdeleteWishlistLoading) {
+      try {
+        await deleteWishlist(bookmarkId._id).unwrap()
+        toast.success("Bookmark deleted successfully")
+      } catch (error) {
+        toast.error((error as ErrorResponse).data.message)
+      }
+    }
+  }
+
+  const handleAddReading = async () => {
+    if (!isaddReadingLoading) {
+      try {
+        await addReading({ bookId: book._id, email: user.email }).unwrap()
+        toast.success("Reading added successfully")
+      } catch (error) {
+        toast.error((error as ErrorResponse).data.message)
+      }
+    }
+  }
+
+  const handleDeleteReading = async () => {
+    if (!isdeleteReadingLoading) {
+      try {
+        await deleteReading(readingId._id).unwrap()
+        toast.success("Reading deleted successfully")
+      } catch (error) {
+        toast.error((error as ErrorResponse).data.message)
+      }
     }
   }
 
@@ -53,19 +95,28 @@ export default function Book(param: Param) {
               <Text weight={500}>{book.title}</Text>
               <Badge>{book.email}</Badge>
             </Flex>
-            {auth ? (
-              bookmarkId ? (
-                <IconBookmarkMinus
-                  cursor={"pointer"}
-                  onClick={handleDeleteBookmark}
-                />
-              ) : (
-                <IconBookmarkPlus
-                  cursor={"pointer"}
-                  onClick={handleAddBookmark}
-                />
-              )
-            ) : null}
+            <Flex gap={8}>
+              {auth ? (
+                bookmarkId ? (
+                  <IconBookmarkMinus
+                    cursor={"pointer"}
+                    onClick={handleDeleteBookmark}
+                  />
+                ) : (
+                  <IconBookmarkPlus
+                    cursor={"pointer"}
+                    onClick={handleAddBookmark}
+                  />
+                )
+              ) : null}
+              {auth ? (
+                !readingId ? (
+                  <IconBook2 cursor={"pointer"} onClick={handleAddReading} />
+                ) : (
+                  <IconBook cursor={"pointer"} onClick={handleDeleteReading} />
+                )
+              ) : null}
+            </Flex>
           </Flex>
         </Group>
